more temporary code
- Modified the draw code to use the new spritebatch test instead of a
single tile
- Color[] textureBackgroundColors - Two default purple-tinted colors
used as the main background texture for tilesheets. Used with the helper
function 'RemoveColors' and the texture that you want to be transparent.
- Added 3 helper functions
* ColorMatch(Color a, Color b, bool checkTransparency = false) - Simply
returns if the two colors match their RGB (and A, if you add true as a
third statement) components
* ReplaceColor(ref Texture2D texture, Color find, Color replace) - Goes
through the entire texture and finds any of the 'find' color, and
replaces it with the 'replace' color. The new data is then set back into
the original texture.
* RemoveColors(ref Texture2D texture, Color[] colors) - Goes through the
entire texture, and replaces any instance that it finds in the list of
colors with a new Color(0, 0, 0, 0). Primary use is to remove the
background for spritesheets using the above non-temporary Color[]
- Using the above helper functions, removes gridded background from
spritesheet